+++ tests/Fixtures/Annotation/BadControllerArgument.php
@@ -3,12 +3,13 @@
 namespace Jungi\FrameworkExtraBundle\Tests\Fixtures\Annotation;
 
 use Jungi\FrameworkExtraBundle\Annotation\QueryParam;
+use Jungi\FrameworkExtraBundle\Annotation\RequestParam;
 
 class BadControllerArgument
 {
     /**
      * @QueryParam("foo")
-     * @QueryParam("foo")
+     * @RequestParam("foo")
      */
     public function bad($foo)
     {