+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - internal `Jungi\FrameworkExtraBundle\Annotation\NamedValueArgument`.
 
 ### Fixed
+- Detecting duplicated annotations on argument by `RegisterControllerAnnotationLocatorsPass`.
 - Deprecation of ReflectionParameter::isArray() in `RegisterControllerAnnotationLocatorsPass`.
 
 ## [1.1.0] - 2020-09-11